change stm32_snail

+++ stm32_snail/common_stm32/grbl/can_stepper.c
@@ -422,7 +422,7 @@ void st_prep_buffer()
     float step_dist_remaining = prep.step_per_mm*mm_remaining; // Convert mm_remaining to steps
     float n_steps_remaining = ceil(step_dist_remaining); // Round-up current steps remaining
     float last_n_steps_remaining = ceil(prep.steps_remaining); // Round-up last steps remaining
-    prep_segment->n_step = last_n_steps_remaining-n_steps_remaining; // Compute number of steps to execute.
+    prep_segment->n_step = (uint16_t)(last_n_steps_remaining-n_steps_remaining); // Compute number of steps to execute.
 
     // Bail if we are at the end of a feed hold and don't have a step to execute.
     if (prep_segment->n_step == 0) {
